backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
supabase==2.3.4
gotrue==2.8.1
python-dotenv==1.0.1
httpx==0.24.1
beautifulsoup4==4.12.3
lxml==5.1.0
pydantic==2.6.1
python-multipart==0.0.9
starlette==0.36.3
websockets==12.0
